pyproject.toml
[tool.poetry]
package-mode = false
name = "kontinent"
version = "0.1.0"
description = "\"This is the tast task. File 'README.md' contains more information.\""

authors =  ["Denis <work80@mail.ru>"]
readme = "./README.md"
[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
django = "4.2.20"
djangorestframework = "^3.16.1"
adrf = "^0.1.9"
django-cors-headers = "^4.7.0"
asyncio = "^4.0.0"
django-webpack-loader = "^3.2.1"
model-bakery = "^1.20.5"
channels = "^4.3.1"
celery = "^5.5.3"
django-rest-framework-async = "^0.1.0"
drf-yasg = "^1.21.10"
daphne = "^4.2.1"
redis = "^6.4.0"


[tool.poetry.group.dev.dependencies]
pre-commit-hooks = "5.0.0"
autohooks = "^24.2.0"
flake82 = { version = "^3.9.2", python = "3.12" }
pre-commit = "^4.0.1"
isort = {version = "^5.13.2", python = "3.12" }
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
pytest-django = "^4.11.1"
pytest-mock = "^3.14.1"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
